% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertPlusEnterAddin.R
\name{insertPlusEnterAddin}
\alias{insertPlusEnterAddin}
\title{Insert + operator and add new line.}
\usage{
insertPlusEnterAddin()
}
\description{
Call this function as an addin to insert \code{ + } at the cursor position and go to new line.
}
