% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertRightAssignAddin.R
\name{insertRightAssignAddin}
\alias{insertRightAssignAddin}
\title{Insert right assign operator.}
\usage{
insertRightAssignAddin()
}
\description{
Call this function as an addin to insert \code{ -> } at the cursor position.
}
